refactor(discord): move runtime orphan recovery helpers
